pyproject.toml
[project]
name = "coinbase-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Eduardo Veloso",email = "eveloso92@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "apache-airflow (>=3.1.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
